Attempt to decouple widgets from glade/Builder
Move to specifying the glade/builder file as a class
variable instead of an init arg.

Add FileBasedWidget and BuilderFileBasedWidget classes in
widgets.py. They are wrappers for gtk.glade.XML() and
or gtk.Builder(). They both will implement from_file()
classmethod contructor, and get_object() and connect_signals(),
since those methods differ between the two and are used heavily.

And a SubmanBaseWidget that takes the place of the previous
widgets.GladeWidget. SubmanBaseWidget has a FileBaseWidget
of sort sort at self.gui, currently it is using BuilderFileBaseWidget.

The main clases for network config dialog, reposgui, factsgui,
preferencesugi, and redeem gui are now subclasses
of SubmanBaseWidget, and use the glade/builder related
convience methods, including the self* names that are
defined in SubmanBaseWidget.widget_names and setup with
pull_widgets(). This eliminates many calls to get_widget/get_object.