Use get_toplevel in place of get_parent_window()
Previous api returned the gdkWindow, which used to
have the associated GtkWindow in it's user data, but
that is no longer the case.

Switch to using widget.get_toplevel() instead, and
adding the approriate is_toplevel() checks as Gtk
docs recomended. get_toplevel() will return the
first GtkWindow in the widgets ancestry, but that
may not be the toplevel GtkWindow in some cases.

stylish cleanup