Fix storage.Mapped*Store's init.
Need to call the Gtk.*Store.__init__ with the list of
column type maps so the Gtk.ListStore is created with
the correct number of columns.

Note: The constructor expects a list of the column types
in _order_. We use the type_map.values(), which is not
in any specific order. Likely need to created a new list
of column types based on looking up the index in self.type_index.