Mapped*Store works now, reenable it.
mysubstab's list widget uses a str type for the column that
shows quantity and serial number, with the values being cast.

Fix test_mysubstab to expect that.

Those columns could be changed to expect int/long types, but
atm we don't have a CellRenderer set for that and we do some
minor tweaking in out text CellRenderer those columns use.